Add method option to hess_log_r
Now, hess_log_r is a bit smarter about your inputs. It allows for an input of a System F, or a Vector of expressions. It will then compute a PWS. It also has an optional method argument that allows you to compute using one of the other two methods we have discussed -- although these are NOT implemented yet.