Add in many slices method + actual hessian
The _many_slices is verifiably broken but is likely close to working. Would like to talk through it with someone.

Also added in a method to take in a known discriminant and compute the hessian of the log directly to compare our results against. The off_diag method seems to work  after a small sign fix.