Add monodromy-free method for `routing_points`
We added a new method for computing the critical points that solves a large polynomial system (using a polyhedral start system) instead of applying monodromy. This gives the correct critical points for the quadratic discriminant, but the mixed volume is very large for the cubic discriminant.

We also noticed that group action is not applicable for the system used for the monodromy approach, since the V parameters break the symmetry.

Co-authored-by: John Cobb <jdcobb3@gmail.com>
Co-authored-by: David K. Johnson <dave-k-johnson@users.noreply.github.com>