Fix bug in construction of lifted lines
When we run monodromy, it seems like we sometimes obtain values of the type `HomotopyContinuation.DoubleDouble.DoubleF64` (aka `ComplexDF64`), which is not compatible with the `LinearSubspace` constructor (which we use in `lifted_line`), and therefore causes an error.

I tried adding a conversion to `ComplexF64` in `LinearSubspace`, which seems to fix the bug for now.