add host-range expansion example
This system models two competing species and a parasite which affects one of the species. The hypersurface in question in this model describes stability boundaries of steady-state solutions and is given by the Routh-Hurwitz criterion. Next steps would be to include the 4-D model mentioned in the paper linked in the file.